backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# AI - Speech to Text (gunakan faster-whisper saja, lebih stabil)
faster-whisper==1.0.0

# AI - Computer Vision & Face Analysis (versi lebih stabil)
opencv-python==4.8.1.78
deepface==0.0.79
fer==22.5.1

# AI - Speech Emotion Recognition
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4

# AI - NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
tokenizers==0.15.0

# AI - Model Optimization (optional, bisa diinstall nanti)
# onnx==1.15.0
# onnxruntime==1.17.0

# Report Generation
reportlab==4.0.9
jinja2==3.1.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
pillow==10.2.0
python-dateutil==2.8.2
